vfio: Rename to vfio-ioctls
