+++ pci/Cargo.toml
@@ -10,7 +10,7 @@ byteorder = "1.3.2"
 devices = { path = "../devices" }
 kvm-bindings = "0.1.1"
 kvm-ioctls = { git = "https://github.com/rust-vmm/kvm-ioctls", branch = "master" }
-vfio = { path = "../vfio" }
+vfio-ioctls = { path = "../vfio-ioctls" }
 vfio-bindings = "0.1.0"
 libc = "0.2.60"
 log = "0.4.8"
