+++ pci/src/vfio.rs
@@ -22,8 +22,8 @@ use std::os::unix::io::AsRawFd;
 use std::ptr::null_mut;
 use std::sync::Arc;
 use std::{fmt, io, result};
-use vfio::{vec_with_array_field, VfioDevice};
 use vfio_bindings::bindings::vfio::*;
+use vfio_ioctls::{vec_with_array_field, VfioDevice};
 use vm_allocator::SystemAllocator;
 use vm_memory::{Address, GuestAddress, GuestUsize};
 use vmm_sys_util::eventfd::EventFd;
