+++ vmm/src/vm.rs
@@ -18,7 +18,6 @@ extern crate linux_loader;
 extern crate net_util;
 extern crate signal_hook;
 #[cfg(feature = "pci_support")]
-extern crate vfio;
 extern crate vm_allocator;
 extern crate vm_memory;
 extern crate vm_virtio;
